T4 SendObject - more debug and bail on empty packets
Added some more debug code with more outputs.

Also found that at end if we did not receive all of the data, we will get empty packets with 0 bytes of data.  Right now bailing when 2 in row in case there are other cases of empty (have not hit that yet).

But then returning back incomplete transfer to host appears to at least keep MTP up and running.