T4.x Handle when running at full speed and not at High Speed
When running at a lower speed, the size of messages also changes from 512 to 64 bytes.

The code was assuming 512 bytes which would crash things when for example you tried to send an object to storage on T4.  It would walk off of end of received buffer and use garbage.

Already had code in core USB to know which version of the interface we bound with and as such the RX and TX size.  So at startup (begin) we ask for the rx and tx size and then use it instead of the hard coded constant size