Updates for CopyObject and MoveObject
This is a combine of multiple commits.

Move/copy check for 0 passed in whichs says use root of destination storage

Uses some of the functions used of SendObject to create the new destination.
This helps when for example the destination directory already exists.

Currently this code will try to overwrite the contents of any of the files
that already existed in the destination with the contentents of the source
files

It will fail if destination exists and not of same type.  That is suppose the
source directory contains a file named temp and the destination already has
a directory named temp.

The changes also tries to reduce how much stack space is used to complete
the operations. The previous code for example the top level operation may
have something like 6 Record objects and if it calls of to the move/copy
directory code each recursion may add similar number of these structures
plus a source and destination pathname object.

Also there is a start for detecting some errors like we ran out of disk
space and the writes fail.  Also work to propagate these errors back
up the call chain.

Fixed CopyObjects for T3.x, we were not properly returning the newly
created object handle and the host would error out.

Cleaned up debug output in storage code, put  in macro DBGPrintf that you did not have to put #if around...