Copy works on t3.6 plus debug outputs cleaned up.
Found copyObject was not working on t3.6 as we did not set the return packet size large enough to return the new object id.

Cleaned up debug output in storage code, put  in macro DBGPrintf that you did not have to put #if around...

Start experimenting with trying to detect more condtions, like write failures.