Make storage part of MTPD plus rename and have explicit one
A simple pass to maybe make example sketches a little cleaner.

Renamed MTPD class to MTP_class like most of the usb classes are of names like:
usb_keyboard_class.

Defined  a standard object: MTP_class MTP

Removed MTPStorage from being a top level object and instead an object contained withing the MTP object.
Moved the Record structure to be contained within the MTPStorage class
Also made the constructor to be constexpr

Added a few wrappers  to have top level MTP.addFileSystem that calls through to the storage one.

added storage() to return pointer to it...

Note: the storage() method may be temporary as we should provide top level object methods for doing
all of the standard functionality.  We may decide to remove it or only use it for non-standard things
like debug functions for dumping the current indexl list.

created keywords.txt file
Updated library.properties file to only include the main header.  This may also be temporary
as this may be automaticaly included when a proper USB type is choosen.

Updated a few of the sketches to use these updates:
	Update Example_2_simple_t41_qflash
	Update Example_3_simple_SD.ino
    Update ssimple_t41_qflash sketch to use this updated stuff.
    Added Probably temporary - added my hardware fs class sketch