Storage Records into 2K blocks compressed
Changed the storage code to not read and write one Record at a time:
Except left code for low memory machines.

Instead you can define how many 2K blocks to hold into memory, and as the
code appends on new records to the storage list, it adds it to current
last block if it will fit with some slop.

I encode which block and which slot in the block into the object ID, that
is the bottom 6 bits is the slot in the block (so max of 64) and currently
the upper 10 bits is the id is the block #.  Note I currently reduced the
ids to only define 16 bits, This reduced the size of each record by about 8 bytes

My experimenting I have been seeing something like 50+ records fit, which depends
on the file name sizes.

There is slop left into each records to allow filenames to be renamed with longer
names.  Still need to add the code to move the data to make room...

Update: It now has code to handle most renames.

Some other changes:

Two defines for filename and path sizes plus messages if fail to read/write storage
The messages are now currently defined to output even if the #define DEBUG
is 0

Changed MAX_FILENAME_LEN to MTP_MAX_FILENAME_LEN
added MTP_MAX_PATH_LEN

Storage:  Creating new items reduce output if DEBUG=0

Previously when you do a do something that creates a new object,
like copy files from pc to teensy,  the create object code was always printing
out the parent record and new record and then dumping the complete index list.

Now Reduced that debug output
This caused too much output, plus interfered with  seeing how many times the cache code had to swap which
buffers were in memory.  Also could not see some error/warnings as the tree was lost in forest.