MTP_Storage.cpp - hopefully fix cache corruption issue.
Found that at least in the case of using SD for cache file, the FS would not work with us doing a seek beyond the end of the file.

Updated the code to avoid this, by  if I needed to output a new cache record beyond the farthest one so far, I would write out the intermediate ones as well.

I also did a little code cleanup and removed trailing spaces on lines.
Plus whe DEBUG 0  I turned off all of the printing except for the error condition ones like a read or write failed