Add a Memory File approach to Index File...
We have run into cases of where the MTP code won't work if for example your
MTP storage list, starts with SD slot and there is no SD in the slot... Or SD is marked
read only.

Right now the code has a way to use the internal memory FS if the user specifies to use the storage index of -2...
But in addition to this, the code will automatically choose this if it fails to open up the storage index file
on the selected (Default first) storage location that was added to MTP

Fix maybe an index issue, by clearing the count of how many records have been stored, when the index file is reopened.
plus change to setting if we should send the reset instead of looking at command which

Fixed a few places which were not building on T3.2