Experiment - detect SD Media code without sketch needing to do it manually
This is an experiment to see if we detect that the FS added to MTP list is an SD object, and if so if we can then at times detect media present or not.

I am doing here by function overload.  Where if it is an SDClass object it will set flag saying it is of type SD and then we have the  Storage Loop function at times check for media present.

It is working but still some issues to resolve. Like when Media is removed , I detect it, but not showing up as RED on MTP... and if I insert different disk, it detects it, and
MTP is not showing it, unless I hit refresh...