Rearrange the callback code
Moved Storage::loop back into storage.

Added to storage a method to set a loop callback by type...

The loop code sees that it is of that type and calls.

Added file with SD callback, with CPP class whose constructor calls the register, and it only should create the object of SD.h was included in the build.